Extract focus helper and sanitize css
